teht4_21

+++ teht4/controllers/blogs.js
@@ -32,13 +32,31 @@ blogsRouter.get('/:id', async (request, response) => {
 blogsRouter.delete('/:id', async (request, response) => {
   try
   {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+    if (!request.token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+    const blog = await Blog.findById(request.params.id)
+    if(blog.user.toString() !== decodedToken.id.toString())
+    {
+      console.log(blog.user.toString())
+      console.log(decodedToken.toString())
+      return response.status(401).json({ error: 'token invalid' })
+    }
+
     await Blog.findByIdAndRemove(request.params.id)
 
     response.status(204).end()
 
   } catch (exception) {
     console.log(exception)
-    response.status(400).send({ error: 'malformatted id' })
+    if (exception.name === 'JsonWebTokenError' ) {
+      response.status(401).json({ error: exception.message })
+    } else {
+      console.log(exception)
+      response.status(500).json({ error: 'something went wrong...' })
+    }
   }
 })
